+++ pytraj/_cyutils.pyx
@@ -1,8 +1,12 @@
 cimport cython
 from cython cimport view
+from .Frame cimport Frame, _Frame
+from .core.Box cimport _Box, Box
+
 
 __all__ = ['_fast_count', 'get_positive_idx', '_int_array1d_like_to_memview',
-           '_int_array2d_like_to_memview', ]
+           '_int_array2d_like_to_memview',
+           '_fast_iterptr', '_fast_iterptr_withbox',]
 
 def _fast_count(cython.integral[:] values, int target):
     cdef int i
@@ -61,3 +65,31 @@ def _int_array1d_like_to_memview(mylist):
             arr0[i] = mylist[i]
         int_view = arr0
     return int_view
+
+def _fast_iterptr(double[:, :, :] xyz, int n_atoms, indices):
+    '''noxbox
+    '''
+    cdef int i
+    cdef int n_frames = xyz.shape[0]
+
+    # just create a pointer
+    cdef Frame frame = Frame(n_atoms, xyz[0], _as_ptr=True)
+    cdef _Frame _frame
+
+    for i in indices:
+        frame.thisptr.SetXptr(n_atoms, &xyz[i, 0, 0])
+        yield frame
+
+def _fast_iterptr_withbox(double[:, :, :] xyz, double[:, :] boxes, int n_atoms, indices):
+    # withbox
+    cdef int i
+    cdef int n_frames = xyz.shape[0]
+
+    # just create a pointer
+    cdef Frame frame = Frame(n_atoms, xyz[0], _as_ptr=True)
+    cdef _Frame _frame
+
+    for i in indices:
+        frame.thisptr.SetXptr(n_atoms, &xyz[i, 0, 0])
+        frame.thisptr.SetBox(_Box(&boxes[i, 0]))
+        yield frame
